# docs/openapi/api.yaml
openapi: 3.1.0
info:
  title: solana-tx-p2p
  description: Solana Transaction Peer-to-peer Node
  license:
    name: ""
  version: 0.0.1
paths:
  /api/v1/peer/discovery:
    post:
      tags:
        - peer::v1
      summary: Discovery peers
      operationId: discovery
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "500":
          description: Internal server error
  /api/v1/peer/relayed-transaction:
    post:
      tags:
        - peer::v1
      summary: List relayed transactions
      operationId: list_relayed_transactions
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "500":
          description: Internal server error
  /api/v1/peer/relayed-transaction/{signature}:
    post:
      tags:
        - peer::v1
      summary: Get relayed transaction
      operationId: get_relayed_transaction
      parameters:
        - name: signature
          in: path
          description: Signature of transaction
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDetail"
        "500":
          description: Internal server error
  /api/v1/peer/signed-message:
    post:
      tags:
        - peer::v1
      summary: List signed messages
      operationId: list_signed_messages
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "500":
          description: Internal server error
components:
  schemas:
    CompiledInstruction:
      type: object
      description: Workaround for `OpenAPI` docs
      required:
        - program_id_index
        - accounts
        - data
      properties:
        accounts:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: |-
            Ordered indices into the transaction keys array indicating which
            accounts to pass to the program.
        data:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: The program input data.
        program_id_index:
          type: integer
          format: int32
          description: |-
            Index into the transaction keys array indicating the program account
            that executes this instruction.
          minimum: 0
    Message:
      type: object
      description: Workaround for `OpenAPI` docs
      required:
        - header
        - account_keys
        - recent_blockhash
        - instructions
      properties:
        account_keys:
          type: array
          items:
            type: string
          description: All the account keys used by this transaction.
        header:
          oneOf:
            - type: object
              description: Workaround for `OpenAPI` docs
              required:
                - num_required_signatures
                - num_readonly_signed_accounts
                - num_readonly_unsigned_accounts
              properties:
                num_readonly_signed_accounts:
                  type: integer
                  format: int32
                  description: |-
                    The last `num_readonly_signed_accounts` of the signed keys are read-only
                    accounts.
                  minimum: 0
                num_readonly_unsigned_accounts:
                  type: integer
                  format: int32
                  description: |-
                    The last `num_readonly_unsigned_accounts` of the unsigned keys are
                    read-only accounts.
                  minimum: 0
                num_required_signatures:
                  type: integer
                  format: int32
                  description: |-
                    The number of signatures required for this message to be considered
                    valid. The signers of those signatures must match the first
                    `num_required_signatures` of [`Message::account_keys`].
                  minimum: 0
          description: The message header, identifying signed and read-only `account_keys`.
        instructions:
          type: array
          items:
            type: object
            description: Workaround for `OpenAPI` docs
            required:
              - program_id_index
              - accounts
              - data
            properties:
              accounts:
                type: array
                items:
                  type: integer
                  format: int32
                  minimum: 0
                description: |-
                  Ordered indices into the transaction keys array indicating which
                  accounts to pass to the program.
              data:
                type: array
                items:
                  type: integer
                  format: int32
                  minimum: 0
                description: The program input data.
              program_id_index:
                type: integer
                format: int32
                description: |-
                  Index into the transaction keys array indicating the program account
                  that executes this instruction.
                minimum: 0
          description: |-
            Programs that will be executed in sequence and committed in one atomic
            transaction if all succeed.
        recent_blockhash:
          type: string
          description: The id of a recent ledger entry.
    MessageHeader:
      type: object
      description: Workaround for `OpenAPI` docs
      required:
        - num_required_signatures
        - num_readonly_signed_accounts
        - num_readonly_unsigned_accounts
      properties:
        num_readonly_signed_accounts:
          type: integer
          format: int32
          description: |-
            The last `num_readonly_signed_accounts` of the signed keys are read-only
            accounts.
          minimum: 0
        num_readonly_unsigned_accounts:
          type: integer
          format: int32
          description: |-
            The last `num_readonly_unsigned_accounts` of the unsigned keys are
            read-only accounts.
          minimum: 0
        num_required_signatures:
          type: integer
          format: int32
          description: |-
            The number of signatures required for this message to be considered
            valid. The signers of those signatures must match the first
            `num_required_signatures` of [`Message::account_keys`].
          minimum: 0
    Transaction:
      type: object
      description: Workaround for `OpenAPI` docs
      required:
        - signatures
        - message
      properties:
        message:
          oneOf:
            - type: object
              description: Workaround for `OpenAPI` docs
              required:
                - header
                - account_keys
                - recent_blockhash
                - instructions
              properties:
                account_keys:
                  type: array
                  items:
                    type: string
                  description: All the account keys used by this transaction.
                header:
                  oneOf:
                    - type: object
                      description: Workaround for `OpenAPI` docs
                      required:
                        - num_required_signatures
                        - num_readonly_signed_accounts
                        - num_readonly_unsigned_accounts
                      properties:
                        num_readonly_signed_accounts:
                          type: integer
                          format: int32
                          description: |-
                            The last `num_readonly_signed_accounts` of the signed keys are read-only
                            accounts.
                          minimum: 0
                        num_readonly_unsigned_accounts:
                          type: integer
                          format: int32
                          description: |-
                            The last `num_readonly_unsigned_accounts` of the unsigned keys are
                            read-only accounts.
                          minimum: 0
                        num_required_signatures:
                          type: integer
                          format: int32
                          description: |-
                            The number of signatures required for this message to be considered
                            valid. The signers of those signatures must match the first
                            `num_required_signatures` of [`Message::account_keys`].
                          minimum: 0
                  description: The message header, identifying signed and read-only `account_keys`.
                instructions:
                  type: array
                  items:
                    type: object
                    description: Workaround for `OpenAPI` docs
                    required:
                      - program_id_index
                      - accounts
                      - data
                    properties:
                      accounts:
                        type: array
                        items:
                          type: integer
                          format: int32
                          minimum: 0
                        description: |-
                          Ordered indices into the transaction keys array indicating which
                          accounts to pass to the program.
                      data:
                        type: array
                        items:
                          type: integer
                          format: int32
                          minimum: 0
                        description: The program input data.
                      program_id_index:
                        type: integer
                        format: int32
                        description: |-
                          Index into the transaction keys array indicating the program account
                          that executes this instruction.
                        minimum: 0
                  description: |-
                    Programs that will be executed in sequence and committed in one atomic
                    transaction if all succeed.
                recent_blockhash:
                  type: string
                  description: The id of a recent ledger entry.
          description: The message to sign.
        signatures:
          type: array
          items:
            type: string
          description: |-
            A set of signatures of a serialized [`Message`], signed by the first
            keys of the `Message`'s [`account_keys`], where the number of signatures
            is equal to [`num_required_signatures`] of the `Message`'s
            [`MessageHeader`].

            [`account_keys`]: Message::account_keys
            [`MessageHeader`]: crate::message::MessageHeader
            [`num_required_signatures`]: crate::message::MessageHeader::num_required_signatures
    TransactionDetail:
      type: object
      required:
        - signatures
        - accountKeys
        - signers
        - logMessages
      properties:
        accountKeys:
          type: array
          items:
            type: string
          description: base58 encoded string of account key
        logMessages:
          type: array
          items:
            type: string
        signatures:
          type: array
          items:
            type: string
          description: base58 encoded string of signature
        signers:
          type: array
          items:
            type: string
          description: base58 encoded string of signer `PeerId`